Learning/04_dependencyInjection_SQLmodel_EnvironmentConfig/pyproject.toml
[project]
name = "04-dependencyinjection-sqlmodel-environmentconfig"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "config>=0.5.1",
    "fastapi>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
]
